+++ utils/index.js
@@ -6,7 +6,7 @@ export const checkPermission = (data, role) => {
 }
 
 export const StripHTMLTags = (value) => {
-  return `${value}`.replace(/(<([^>]+)>)/gi, '').replace(/&nbsp;/gi, ' ')
+  return `${value}`.replace(/(<([^>]+)>)/gi, '').replace(/&nbsp;/gi, ' ');
 }
 
 export const getTimestamp = (value) => {
@@ -15,6 +15,10 @@ export const getTimestamp = (value) => {
     if (diff < 1 * 60 * 1000) return 'Just now';
     return `${format(diff, true)} ago`;
   }, [value]);
+}
+
+export const parseDataPage = (data, key = 'data') => {
+  return Array.isArray(data) ? data.reduce((acc, val) => [...acc, ...val[key]], []) : [];
 }
 
 export default {};